+++ src/main/kotlin/com/coder/toolbox/settings/ReadOnlyCoderSettings.kt
@@ -35,10 +35,10 @@ interface ReadOnlyCoderSettings {
     val binarySource: String?
 
     /**
-     * Directories are created here that store the CLI for each domain to which
-     * the plugin connects.   Defaults to the data directory.
+     * An absolute path to a local directly where the CLI will be downloaded. If [enableDownloads] is true
+     * then this setting can point to the CLI file locally. Defaults to the data directory.
      */
-    val binaryDirectory: String?
+    val binaryDestination: String?
 
     /**
      * Controls whether we verify the cli signature
@@ -61,18 +61,13 @@ interface ReadOnlyCoderSettings {
     val defaultCliBinaryNameByOsAndArch: String
 
     /**
-     * Configurable CLI binary name with extension, dependent on OS and arch
-     */
-    val binaryName: String
-
-    /**
      * Default CLI signature name based on OS and architecture
      */
     val defaultSignatureNameByOsAndArch: String
 
     /**
      * Where to save plugin data like the Coder binary (if not configured with
-     * binaryDirectory) and the deployment URL and session token.
+     * binaryDestination) and the deployment URL and session token.
      */
     val dataDirectory: String?